Use base class for BC trainers
