+++ ml-agents/mlagents/trainers/bc/offline_trainer.py
@@ -10,6 +10,7 @@ import tensorflow as tf
 
 from mlagents.envs import AllBrainInfo
 from mlagents.trainers.bc.policy import BCPolicy
+from mlagents.trainers.bc.trainer import BCTrainer
 from mlagents.trainers.buffer import Buffer
 from mlagents.trainers.demo_loader import DemonstrationLoader
 from mlagents.trainers.trainer import UnityTrainerException, Trainer
@@ -17,8 +18,8 @@ from mlagents.trainers.trainer import UnityTrainerException, Trainer
 logger = logging.getLogger("mlagents.trainers")
 
 
-class OfflineBCTrainer(Trainer):
-    """The OfflineBCTrainer is an implemntation of Offline Behavioral Cloning."""
+class OfflineBCTrainer(BCTrainer):
+    """The OfflineBCTrainer is an implementation of Offline Behavioral Cloning."""
 
     def __init__(self, brain, trainer_parameters, training, load, seed, run_id):
         """
@@ -29,40 +30,26 @@ class OfflineBCTrainer(Trainer):
         :param seed: The seed the model will be initialized with
         :param run_id: The The identifier of the current run
         """
+        super(OfflineBCTrainer, self).__init__(brain, trainer_parameters, training, load, seed, run_id)
+        self.TRAINER_NAME = "Offline Behavioral Cloning"
+
         self.param_keys = ['batch_size', 'summary_freq', 'max_steps',
                            'batches_per_epoch', 'use_recurrent',
                            'hidden_units', 'learning_rate', 'num_layers',
                            'sequence_length', 'memory_size', 'model_path',
                            'demo_path']
 
-        for k in self.param_keys:
-            if k not in trainer_parameters:
-                raise UnityTrainerException(
-                    "The hyperparameter {0} could not be found for the Imitation trainer of "
-                    "brain {1}.".format(k, brain.brain_name))
-
-        super(OfflineBCTrainer, self).__init__(brain, trainer_parameters, training, run_id)
-
+        self.check_param_keys(self.TRAINER_NAME, self.param_keys)
         self.policy = BCPolicy(seed, brain, trainer_parameters, load)
-        self.brain_name = brain.brain_name
         self.batches_per_epoch = trainer_parameters['batches_per_epoch']
         self.n_sequences = max(int(trainer_parameters['batch_size'] / self.policy.sequence_length),
                                1)
-        self.cumulative_rewards = {}
-        self.episode_steps = {}
-        self.stats = {'losses': [], 'episode_length': [], 'cumulative_reward': []}
-
-        self.training_buffer = Buffer()
-        self.summary_path = trainer_parameters['summary_path']
-        if not os.path.exists(self.summary_path):
-            os.makedirs(self.summary_path)
-
-        self.summary_writer = tf.summary.FileWriter(self.summary_path)
 
         brain_params, self.demo_buffer = DemonstrationLoader.load(trainer_parameters['demo_path'],
                                                                   self.brain_name,
                                                                   self.policy.sequence_length)
 
+        print(len(self.demo_buffer.update_buffer['actions']), self.n_sequences)
         if brain.__dict__ != brain_params.__dict__:
             raise UnityTrainerException("The provided demonstration is not compatible with the "
                                         "brain being used for performance evaluation.")
@@ -71,121 +58,6 @@ class OfflineBCTrainer(Trainer):
         return '''Hyperparameters for the Imitation Trainer of brain {0}: \n{1}'''.format(
             self.brain_name, '\n'.join(
                 ['\t{0}:\t{1}'.format(x, self.trainer_parameters[x]) for x in self.param_keys]))
-
-    @property
-    def parameters(self):
-        """
-        Returns the trainer parameters of the trainer.
-        """
-        return self.trainer_parameters
-
-    @property
-    def get_max_steps(self):
-        """
-        Returns the maximum number of steps. Is used to know when the trainer should be stopped.
-        :return: The maximum number of steps of the trainer
-        """
-        return float(self.trainer_parameters['max_steps'])
-
-    @property
-    def get_step(self):
-        """
-        Returns the number of steps the trainer has performed
-        :return: the step count of the trainer
-        """
-        return self.policy.get_current_step()
-
-    @property
-    def get_last_reward(self):
-        """
-        Returns the last reward the trainer has had
-        :return: the new last reward
-        """
-        if len(self.stats['cumulative_reward']) > 0:
-            return np.mean(self.stats['cumulative_reward'])
-        else:
-            return 0
-
-    def increment_step_and_update_last_reward(self):
-        """
-        Increment the step count of the trainer and Updates the last reward
-        """
-        self.policy.increment_step()
-        return
-
-    def take_action(self, all_brain_info: AllBrainInfo):
-        """
-        Decides actions using policy given current brain info.
-        :param all_brain_info: AllBrainInfo from environment.
-        :return: a tuple containing action, memories, values and an object
-        to be passed to add experiences
-        """
-        if len(all_brain_info[self.brain_name].agents) == 0:
-            return [], [], [], None, None
-
-        agent_brain = all_brain_info[self.brain_name]
-        run_out = self.policy.evaluate(agent_brain)
-        if self.policy.use_recurrent:
-            return run_out['action'], run_out['memory_out'], None, None, None
-        else:
-            return run_out['action'], None, None, None, None
-
-    def add_experiences(self, curr_info: AllBrainInfo, next_info: AllBrainInfo,
-                        take_action_outputs):
-        """
-        Adds experiences to each agent's experience history.
-        :param curr_info: Current AllBrainInfo (Dictionary of current brains and BrainInfos).
-        :param next_info: Next AllBrainInfo (Dictionary of current brains and BrainInfos).
-        :param take_action_outputs: The outputs of the take action method.
-        """
-        info_student = curr_info[self.brain_name]
-        next_info_student = next_info[self.brain_name]
-        for agent_id in info_student.agents:
-            self.training_buffer[agent_id].last_brain_info = info_student
-
-        # Used to collect information about student performance.
-        for agent_id in next_info_student.agents:
-            stored_info_student = self.training_buffer[agent_id].last_brain_info
-            if stored_info_student is None:
-                continue
-            else:
-                next_idx = next_info_student.agents.index(agent_id)
-                if agent_id not in self.cumulative_rewards:
-                    self.cumulative_rewards[agent_id] = 0
-                self.cumulative_rewards[agent_id] += next_info_student.rewards[next_idx]
-                if not next_info_student.local_done[next_idx]:
-                    if agent_id not in self.episode_steps:
-                        self.episode_steps[agent_id] = 0
-                    self.episode_steps[agent_id] += 1
-
-    def process_experiences(self, current_info: AllBrainInfo, next_info: AllBrainInfo):
-        """
-        Checks agent histories for processing condition, and processes them as necessary.
-        Processing involves calculating value and advantage targets for model updating step.
-        :param current_info: Current AllBrainInfo
-        :param next_info: Next AllBrainInfo
-        """
-        info_student = next_info[self.brain_name]
-        for l in range(len(info_student.agents)):
-            if info_student.local_done[l]:
-                agent_id = info_student.agents[l]
-                self.stats['cumulative_reward'].append(
-                    self.cumulative_rewards.get(agent_id, 0))
-                self.stats['episode_length'].append(
-                    self.episode_steps.get(agent_id, 0))
-                self.cumulative_rewards[agent_id] = 0
-                self.episode_steps[agent_id] = 0
-
-    def end_episode(self):
-        """
-        A signal that the Episode has ended. The buffer must be reset. 
-        Get only called when the academy resets.
-        """
-        self.training_buffer.reset_all()
-        for agent_id in self.cumulative_rewards:
-            self.cumulative_rewards[agent_id] = 0
-        for agent_id in self.episode_steps:
-            self.episode_steps[agent_id] = 0
 
     def is_ready_update(self):
         """
