+++ ml-agents/mlagents/trainers/ppo/trainer.py
@@ -29,17 +29,14 @@ class PPOTrainer(Trainer):
         :param seed: The seed the model will be initialized with
         :param run_id: The The identifier of the current run
         """
+        super(PPOTrainer, self).__init__(brain, trainer_parameters, training, run_id)
         self.param_keys = ['batch_size', 'beta', 'buffer_size', 'epsilon', 'gamma', 'hidden_units', 'lambd',
                            'learning_rate', 'max_steps', 'normalize', 'num_epoch', 'num_layers',
                            'time_horizon', 'sequence_length', 'summary_freq', 'use_recurrent',
                            'summary_path', 'memory_size', 'use_curiosity', 'curiosity_strength',
                            'curiosity_enc_size', 'model_path']
-
-        for k in self.param_keys:
-            if k not in trainer_parameters:
-                raise UnityTrainerException("The hyperparameter {0} could not be found for the PPO trainer of "
-                                            "brain {1}.".format(k, brain.brain_name))
-        super(PPOTrainer, self).__init__(brain.brain_name, trainer_parameters, training, run_id)
+        self.TRAINER_NAME = "Proximal Policy Optimization"
+        self.check_param_keys(self.TRAINER_NAME, self.param_keys)
 
         self.use_curiosity = bool(trainer_parameters['use_curiosity'])
